requirements.txt
asgiref==3.4.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
cryptography==37.0.4
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==3.2.18
django-allauth==0.51.0
django-cors-headers==3.10.1
django-environ==0.9.0
django-storages==1.12.3
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
jmespath==0.10.0
oauthlib==3.2.0
Pillow==8.4.0
psycopg2
pycparser==2.21
PyJWT==2.4.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.2.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.1.1
tzdata==2022.2
urllib3==1.26.11
django-oauth-toolkit==2.2.0